Camera device properties and FacadeKit library
Additional properties on the Camera device source have been added to aid
with the identification of Facade devices. In particular, the
"kMagicProperty" is used to identify all Facade-managed devices. Other
properties have been exposed to build on later.

The FacadeKit dynamic library is a C API that will be used in other
language bindings for writing to Facade devices. So far, the FacadeCLI
application is able to use it to write a solid purple color into the
first Facade virtual camera.